Remove name datastructures
No separate symbol resolution stage, as this is unnecessary complicated. This branch is still useful by removing the need for lifetime annotations. Symbol resolution can be easily done during typechecking by adding a list of still-to-resolved types representing the imports to a typechecked module and checking them at some later point. Because we use HM, this should be possible, and might even allow cyclic dependencies (as long as we use one big context, that's TODO)